chore: move source and tests into Vite+ scaffold
Move source files into file-selector/src/ and test files into
file-selector/tests/, adjusting imports for the new layout.
