--- file-selector/tests/index.test.ts
@@ -1,6 +1,0 @@
-import { expect, test } from "vite-plus/test";
-import { fn } from "../src/index.ts";
-
-test("fn", () => {
-    expect(fn()).toBe("Hello, tsdown!");
-});
