--- src/index.ts
@@ -1,2 +1,0 @@
-export {fromEvent} from './file-selector';
-export {FileWithPath} from './file';
